requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==3.6.2
argcomplete==3.2.2
asttokens==2.4.1
attrs==22.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
bokeh==3.3.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deepgram-sdk==3.1.7
diagrams==0.23.4
distro==1.9.0
dnspython==2.3.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.85.2
fastapi-utils==0.2.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
frozenlist==1.4.1
gensim==4.3.2
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.42
graphviz==0.20.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.55
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.24.1
openai==1.14.3
openpyxl==3.0.10
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.2
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.1.0
pluggy==1.0.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==15.0.0
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.3.3
PyNaCl==1.5.0
PySocks==1.7.1
pyspark==3.3.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
scp==0.14.5
selenium==4.17.2
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.3
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.46
stack-data==0.6.3
starlette==0.20.4
streamlit==1.31.1
streamlit-bokeh-events==0.1.2
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typed-ast==1.5.5
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
userpath==1.9.1
uvicorn==0.18.3
validators==0.22.0
verboselogs==1.7
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
